pyproject.toml
[project]
name = "library-management-system"
version = "0.1.0"
description = "Mini Library Management System built with FastAPI"
requires-python = ">=3.10"
dependencies = [
    "fastapi>=0.111",
    "uvicorn[standard]>=0.30",
    "sqlalchemy[asyncio]>=2.0",
    "asyncpg>=0.29",
    "alembic>=1.13",
    "pydantic-settings>=2.2",
    "authlib>=1.3",
    "python-jose[cryptography]>=3.3",
    "httpx>=0.27",
    "passlib[bcrypt]>=1.7",
    "python-multipart>=0.0.9",
    "itsdangerous>=2.1",
    "openai>=1.30",
]

[project.optional-dependencies]
dev = [
    "ruff",
    "black",
    "pytest",
    "pytest-asyncio",
    "pytest-cov",
]

[tool.ruff]
line-length = 100

[tool.ruff.lint]
select = ["E", "W", "F", "I"]
ignore = ["E501"]

[tool.black]
line-length = 100

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["app/tests"]

[tool.setuptools.packages.find]
where = ["."]
include = ["app*"]

[build-system]
requires = ["setuptools>=68"]
build-backend = "setuptools.build_meta"
